* merge-kent-john:
  ui fix
  added support for directions
  ux improvments and fixes
  audio search
  Added tags for searching with messages
  added support for natural language search (text)
  Added audio parsing to search nearby
  started to add support for voice search
  Added searching function with wit.ai
  Added owner field to polygons
  search improvements
  Added owner field to pins
  Fixed errors from merge
  Added wit package
  integrate geocoding
  Added signout and new error handling
  Added string error handling
  Moved authorization for now
  Fixing some imports

+++ backend/db/resolvers/polygon-resolver.js
@@ -1,12 +1,11 @@
 const Polygon = require('../models/polygon-model');
 const Pin = require('../models/pin-model');
-const {isAuthenticated} = require('../../util');
-
+const {isAuthenticated, isAuthorized} = require('../../util');
 
 createPolygon = async function (input, context) {
     let auth = isAuthenticated(context.req);
     if (auth) return auth();
-    const polygonInput = Object.assign({}, input, {user: context.req.session.user});
+    const polygonInput = Object.assign({}, input, {user: context.req.session.user, owner: context.req.session.user.username});
     const polygon = await new Polygon(polygonInput).save();
     return polygon;
 };
@@ -57,6 +56,8 @@ deletePolygon = async function(input, context) {
     let auth = isAuthenticated(context.req);
     if (auth) return auth();
     const polygon = await Polygon.findOne({_id: context.req.params.id}).exec();
+    let perm = isAuthorized(context.req, polygon.user);
+    if (perm) return perm();
     Polygon.deleteOne({_id: polygon._id}).exec();
     return null;
 }
